removed and replaced files to match x64 Version 3 virtual storage model(s)
