+++ src/libgxxjit_interpreter.c
@@ -5,7 +5,7 @@
 #include "../inc/coverage_testassert_debug_mod.h"
 #include "../inc/handling.h"
 #include "../inc/ndebug_testassert_messages.h"
-#include "../inc/powerpc_virtual_storage_model.h"
+#include "../inc/cellbe_harvard_style_cache.h"
 
 #if _WIN32
 #include <memory.h>
