+++ src/main.c
@@ -14,54 +14,6 @@
 #include <stdlib.h>
 #endif // !_WIN32 ||  !__linux__
 
-#include "../inc/amd64_command_buffer_handler.h"
-#include "../inc/amd64_command_ring_buffer.h"
-#include "../inc/components.h"
-#include "../inc/coverage_testassert_debug_mod.h"
-#include "../inc/handling.h"
-#include "../inc/ndebug_testassert_messages.h"
-#include "../inc/powerpc_virtual_storage_model.h"
-
-#ifdef ASSERT_VARGS
-// #define ASSERT_VARGS(_str, ...) // redefined for re-use
-#endif // !ASSERT_VARGS
-// extern ASSERT_VARGS(_str, ...)
-
-uint16_t read_virtual_enumerated_header(
-    struct PpuOffsetLookasideBuffer *address_offset) {
-  uint64_t ret_val = 0;
-
-  FILE *ifloop_file, *switchstate_file;
-  DecryptedCommandBufferTree *command_buffer_tree;
-  COVER_TESTASSERT_IF(CURRENT_COLUMN, SHORT_OFFSET, ifloop_file->_cur_column);
-
-  switch (address_offset->rb_status_code) {
-  case SUCCESS:
-    ret_val =
-        (uint64_t)
-            address_offset->ppe_translation_lookaside_buffer.address_offset;
-    assert(address_offset->rb_status_code);
-    assert(SUCCESS);
-
-  default:
-    ret_val = 1;
-  }
-  COVER_TESTASSERT_SWITCHSTATE(
-      CURRENT_COLUMN, SHORT_OFFSET,
-      address_offset->ppe_translation_lookaside_buffer.address_offset,
-      switchstate_file->_vtable_offset);
-
-  return ret_val;
-}
-
-uint16_t
-write_into_enumerated_header(struct PpuOffsetLookasideBuffer *address_offset,
-                             int io_insertions, FILE *buffer_file) {
-  FILE *ifloop_file = buffer_file, *switchstate_file = buffer_file;
-  COVER_TESTASSERT_IF(CURRENT_COLUMN, SHORT_OFFSET, ifloop_file);
-  return 0;
-}
-
 int exec_system_setup() {
   Texture2D dubois_crance_sprite =
       LoadTexture("assets/sprites/DuboisCranceM1A1-Export/spritesheet.png");
