--- src/powerpc_function_offload_hook.c
@@ -1,8 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-typedef struct FunctionOffloadHookId {
-  int bytes_n;
-  char message[];
-} FunctionOffloadHookId;
