--- src/powerpc_spu_command_handler.c
@@ -1,181 +1,0 @@
-#define __va_args_sentinel__(M, ...)                                           \
-  {                                                                            \
-    errno = __LINE__;                                                          \
-    __FILE__;                                                                  \
-  }
-
-// #include <openssl/sha.h>
-// int SHA256_Init(SHA256_CTX *c);
-// int SHA256_Update(SHA256_CTX *c, const void *data, size_t len);
-// int SHA384_Final(unsigned char *md, SHA256_CTX *c);
-
-#include "../inc/powerpc_spu_command_handler.h"
-#include "../inc/amd64_command_ring_buffer.h"
-
-#include <errno.h>
-#include <fcntl.h>
-#include <inttypes.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <sys/stat.h>
-#include <io.h> // windows equal to unistd
-
-#ifndef CHUNK_BUFFER_SLICE
-#define CHUNK_BUFFER_SLICE 1024
-#endif
-
-const int _va_args_filedescs_sentinel(int argc, char *argv[]) {
-  int input_filedesc, output_filedesc, open_flags;
-  char buffers[CHUNK_BUFFER_SLICE];
-  int ret_val = 0;
-  FILE *file;
-
-  if (argc != 3 || strcmp(argv[1], "--help") == 0)
-    ret_val = 1;
-
-  // NOTE: open(...) is an <fcntl.h> function
-  int errno_t = fopen_s(&file, argv[1], O_RDONLY); // open is deprecated. Use: _sopen_s
-
-  if (fclose(file) == -1)
-    ret_val = -1;
-
-  return ret_val;
-}
-
-inline const uint32_t init_system_startup_prop(int context_flag_nu_x0) {
-  context_flag_nu_x0 = 0x0000; // equivalent to SUCCESS enumerated value
-  if (context_flag_nu_x0 != SUCCESS)
-    context_flag_nu_x0 ^=
-        RB_MODE_EXITED; // is-equal-to: context_flag_nu_x0 ^ RB_MODE_EXITED
-  return context_flag_nu_x0;
-}
-
-const int setup_startup_system(int (*outlined_startup_context)()) {
-  int ret_val = 0;
-
-  // memcpy(inlined_destiny, inlined_source, sizeof(InlinedEntity));
-
-  enum RbStatusCodes rb_status_code;
-
-  switch (rb_status_code) {
-  case SUCCESS:
-    ret_val = 0x0000;
-  default:
-    break;
-  }
-
-  return ret_val;
-}
-
-const int fault_decrypted_entity_proc(enum RbStatusCodes rb_status_code) {
-  int ret_val = 0;
-
-  switch (rb_status_code) {
-  case SUCCESS:
-    ret_val = 0;
-  default:
-    if (SUCCESS != 0x0000) { // we want: init_system_startup_prop != SUCCESS
-      ret_val = 1;
-      __va_args_sentinel__(ret_val);
-    }
-    break;
-  }
-  return ret_val;
-}
-
-const int *determine_inferred_101_load(CommandRingBuffer *restrict cell_unit) {
-  const float light_weight = 0.0000f;
-  const float heavy_weight = 1.0000f;
-
-  // for (int *i = 0; i < cell_unit->hashkeys->keys; i++)
-  //     for (int *j = 0; j < cell_unit->hashkeys->keys; j++)
-  //     {
-  //         cell_unit->weight = light_weight;
-  //         cell_unit->hashkeys->keys[/*sha256_value*/ *i] = *j;
-  //     }
-
-  return 0;
-} // Scheme 1
-
-inline const int *
-report_101_patron_info(CommandRingBuffer *restrict cell_unit) {
-  determine_inferred_101_load(cell_unit);
-
-  // for (int *k = 0; k < cell_unit->hashkeys->keys; k++)
-  // {
-  //     printf("\nNetwork peer => Hashkey := \n\t%d",
-  //            cell_unit->hashkeys->keys[*k]);
-  // }
-  return 0;
-} // Scheme 2
-
-inline const int *
-many_many_directory_compute(CommandRingBuffer *restrict cell_units) {
-  report_101_patron_info(cell_units);
-  for (int i = 0; i < sizeof(CommandRingBuffer);
-       i++) // change sizeof(ArbitraryCellUnit) to ...
-    for (int j = 0; j < sizeof(CommandRingBuffer);
-         j++) { // cell.arena_state.data_size
-      CommandRingBuffer network_x_peers = cell_units[i];
-      CommandRingBuffer network_y_peers = cell_units[j];
-    }
-  int *key = 0;
-
-  //   for (int *i = 0; i || network_x_peers++;
-  //        ++i) // what if this is from a file data-size?
-  //     for (int *j = 0; j || network_y_peers--;
-  //          --j) { // come from EOF and go-backwards
-  //       cell_units[*key]->weight = 0.0000f;
-  //       cell_units[*key]->hashkeys->keys[/*sha256_value*/ *i] = *j;
-  //     }
-
-  return 0;
-} // Scheme 3
-
-inline struct PhantomMarker load_balance_pow2() {
-  struct PhantomMarker *phantom_marker;
-  struct DecryptedCommandBufferCursor *restrict decrypted_cursor;
-  struct DecryptedCommandBufferTree *restrict buffer_tree;
-
-  struct PhantomMarker ret_val = buffer_tree->phantom_sizes[0][0];
-  for (int i = phantom_marker->cursor_position;
-       i < sizeof(decrypted_cursor->phantom_marker); ++i)
-    for (int j = phantom_marker->read_position;
-         j < sizeof(buffer_tree->phantom_sizes); ++j) {
-      ret_val = buffer_tree->phantom_sizes[i][j];
-      if ((i ^ j) == SUCCESS) {
-        int k = i ^ j;
-        ret_val = buffer_tree->phantom_sizes[k][k];
-      }
-    }
-
-  return ret_val;
-}
-
-inline const int *unload_balanced_loads(/*SHA256_CTX *c*/) {
-  // auto context_data = memset(context_destiny, context_source, sizeof(c));
-  // auto user_data = UserData {.destination = context_destiny, .source =
-  // context_source, .user_data = c, .data_size = sizeof(c)};
-  // free(user_data);
-  // free(context_data);
-
-  return 0;
-}
-
-inline const int *insert_balanced_loads(/*SHA256_CTX *c*/) {
-  // auto context_data = memset(context_destiny, context_source, sizeof(c));
-  // auto user_data = UserData {.destination = context_destiny, .source =
-  // context_source, .user_data = c, .data_size = sizeof(c)};
-
-  return 0;
-}
-
-const int *add_outlined_system(int (*context)()) {
-  // struct InlinedEntity *inlined_context;
-  // inlined_context = memset(inlined_context, SUCCESS, sizeof(InlinedEntity));
-
-  load_balance_pow2();
-
-  return 0;
-}
